@Nandaka 2020-06-27
nah, if that the case, then I think only `__config__.loadConfig(path=configfile)` is the one needed to be moved inside `processFanboxArtist()`