@bluerthanever 2020-06-27
well, if **ONLY** `__config__.loadConfig(path=configfile)` is moved inside `processFanboxArtist()`, then for `menu_fanbox_dowload_by_artist_or_creator_id`, the sequence would be:
1. creating a `FanboxArtist` instance with `__br__.fanboxGetCArtistById`, in which `__config__.writeConfig` would be called
2. go into `processFanboxArtist`, and `loadConfig` will be called which i don't think is necessary

and in `menu_fanbox_download_from_artist`, for each `FanboxArtist` instance:
1. ` __br__.fanboxUpdateArtistToken` is called firstly, in which `__config__.writeConfig` would be called
2.  go into `processFanboxArtist`, and `loadConfig` will be called 
this order seems weird to me.... unless you are talking about putting '__config__.loadConfig(path=configfile)` at the end