@Nandaka 2020-06-28
I see, because of the side effect of `__br__.fanboxUpdateArtistToken` causing the confgi to be updated, but there is a risk of overwriting if the config was changed in between.
If that the case, then maybe adding `__config__.loadConfig(path)` before calling `processFanboxArtist()` in ' menu_fanbox_download_by_artist_or_creator_id' will do?