@bluerthanever 2020-06-28
huh? you mean still move `__config__.loadConfig(path=configfile)` inside `processFanboxArtist`, and then add `__config__.loadConfig(path)` before calling `processFanboxArtist` in `menu_fanbox_download_by_artist_or_creator_id`?
so when processing a single id, the order would be `loadConfig`, then `fanboxGetArtistById` in which `writeConfig` will be called, then enters `processArtist`, and `loadConfig` again? **load -> write -> load**
and for supporting or following artists, it's **write -> load**?