@bluerthanever 2020-06-28
Never mind. I just found another place to optimize, I will close this one and change it to the later form.