@Nandaka 2020-06-28
No need to make it so complicated.
```
def menu_fanbox_download_by_artist_or_creator_id(op_is_valid, args):
    __log__.info('Download FANBOX by Artist or Creator ID mode.')
    end_page = 0
    id = ''

    if op_is_valid and len(args) > 0:
        id = args[0]
        if len(args) > 1:
            end_page = args[1]
    else:
        id = input("Artist/Creator ID = ").rstrip("\r")
        end_page = input("Max Page = ").rstrip("\r") or 0

    end_page = int(end_page)

    __config__.loadConfig(path)
    artist = __br__.fanboxGetArtistById(id)    
    processFanboxArtist(artist, end_page)
```