@bluerthanever 2020-06-28
I get what you meant. But changing it like this doesn't give the guy much time to change the configuration, **IF** the refresh token is changed.