@Nandaka 2020-06-28
well, it is a console application. In windows, you can pause the processing by clicking/selecting the console using mouse (if you enable Quick Edit mode), not sure about linux terminal.